requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg2-binary>=2.9.10
celery==5.3.6
redis==5.0.3
requests==2.31.0
django-environ==0.11.2
django-filter==24.1
drf-spectacular
